#91 - broke installation steps into sections
This is not strictly part of the implementation of this issue.  However, I was trying to find some specific steps while working on the issue.  I was annoyed at how much scrutiny was required to find the steps, so I added section headings to make it easier.

The Markdown editor I use (Typora) also updated some formatting automatically.