#108 - resolve `errors` variable conflict
Use of `errors` for list of requests that failed due to error conflicted with `errors` in `APIErrorData`.  Resolved by renaming to `failures`.  (Makes sense, because "failure" is the opposite of "success".)  Also going to find out if IntelliJ IDEA's eslint formatting agrees with Codacy's checks, esp. WRT multiple assignment in braces.