#108 - remove user lookup code
The specifications were unclear and I was not familiar with Canvas API's `sis_login_id:` prefix for referring to users by uniqname in requests rather than a Canvas ID number.  If I'd known that notation was a literal API token rather than pseudocode, it would have saved about a week of development time.