#108 - Got retry experimentation
Added some hooks to log messages at key points in the request lifecycle.  With this code, I see get log lines like `ccm_web     | 2021-08-20T16:42:23.670Z - canvas.service.ts - debug - ♻♻♻♻ beforeRetry ♻♻♻♻ [1]: ECONNRESET`.  So, although Canvas API is returning status code 403, which is not caught by the retry code, it appears that status causes a ECONNRESET error, which is handled.  Unfortunately, it seems that the retry is not successful.  Perhaps the retries are happening too soon.