#291 - removed, so user can accept terms
From Canvas API docs: `user[terms_of_use]` - Whether the user accepts the terms of use. Required if this is a self-registration and this canvas instance requires users to accept the terms (on by default). If this is true, it will mark the user as having accepted the terms of use.