#221 - error handling changes
The error handling code makes some assumptions about the structure of responses from the Canvas API.  In the case of creating users, when there were errors (e.g., the user already exists, etc.) this code didn't recognize it as a proper Canvas error, because it had a different format.  So, added a way for other error message formats to be accepted.