#221 - move swagger link to top
This change isn't strictly related to the issue, but I prefer the Swagger link to appear at the top of the page.  With the app loaded into an iframe and the iframe doesn't fit entirely within the page's display (at least, not in my browser), there's a lot of scrolling within scrolling I need to do when the link is at the bottom of the page.  It's much simpler at the top.  Plus, with the link at the top, we could also add a "DEVELOPMENT MODE" banner there, if we'd like.