#221 - solve some of our invalid email problems
I just noticed the `IsEmail` validator.  It's going to be very helpful to us.  I discovered that if the input contains invalid email addresses, Canvas is OK with that and creates the account if it doesn't exist.  Not ideal.  Even worse, the next step of the process passes the email addresses (including the invalid ones) that need to be invited to Cirrus.  Unfortunately, when Cirrus gets invalid email addresses in the batch input, the entire batch fails, so nobody gets an invitation.  That means accounts were created (and subsequent enrollments were made for them), but no invitations were sent.  I was thinking that it may mean the process would need to be reworked to look everybody up FIRST, attempt to invite that batch of new people, and if that succeeds, then do the account creation.  The order of those steps and the additional number of steps is not desirable.  Now that we can call the entire request invalid if any email addresses are invalid, that should save us from any major changes.  I will test further to verify.